Revert "Add key warning to nested collections"
This heuristic isn't great because it relies on inspecting deep children which aren't guaranteed to be React elements. In particular, this was causing stack overflows in a component we had that used a *DOM node* as children, like `<DOMContainer>{node}</DOMContainer>`.

This reverts commits:
ad2efe57505bf167ecd9b4d30b7cd87e387551e7
0def6b90bcd56342083109a7561caa96eff8db70
1335fe5de3751375fc50660a07a93b517ac885b4
4511a1fd72f3a7ff8acab28156ef57398fd99715
3a17a628a975d983d8a997e60b1ad35261b816e9