Revert "Use the virtual target for change events to avoid restoring controlled state on the real target (#10444)" (#10504)
This reverts the meaningful (src, non-test) part of commit 44460528585ed85f00e331d6d6f237f752206567 since we've reverted the commit it depended on in 6efcad5aca75ec1e254595c8c347bb582bec01be. I don't fully understand the negative implications of leaving this unreverted but it sounds like consensus is that it's safer to revert.

I left the new fixture and verified it works correctly in Chrome 62 Mac, as well as the jest tests passing.