Allow functions to be passed to custom element setters (#23042)
This is part of the new custom element features that were implemented
here:
https://github.com/facebook/react/commit/b4036525657b4ac9e6c11446b0ade499f5d44b20

When a custom element has a setter for a property and passes the `in`
heuristic, the value passed to the react property should be assigned
directly to the custom element's property, regardless of the type of the
value. However, it was discovered that this isn't working with
functions. This patch makes it work with functions.

Fixes https://github.com/facebook/react/issues/23041